Move signaling to more common folder
