+++ src/signals/incident-management/components/Signaling/components/BarGraph/BarGraph.tsx
