+++ src/signals/incident-management/components/Signaling/components/BarGraph/__tests__/BarGraph.test.tsx
