+++ src/signals/incident-management/components/Signaling/components/BarGraph/index.ts
