+++ src/signals/incident-management/components/Signaling/components/GraphDescription/index.ts
