+++ src/signals/incident-management/components/Signaling/components/GraphEmpty/index.ts
