+++ src/signals/incident-management/components/Signaling/index.ts
