+++ src/signals/incident-management/containers/Dashboard/components/index.ts
@@ -1,4 +1,4 @@
 export { AreaChart } from './AreaChart'
 export { BarChart } from './BarChart'
 export { Filter } from './Filter'
-export { default as Signaling } from './Signaling'
+export { default as Signaling } from '../../../components/Signaling'
